Finished chapter 4: Displaying a list
The Tour of Heroes app displays a list of heroes in a Master/Detail view.
The user can select a hero and see that hero's details.
Used *ngFor to display a list.
Used *ngIf to conditionally include or exclude a block of HTML.
Can toggle a CSS style class with a class binding.